# docs-swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Microservice PhucStore V1
  description: |-
    Đây là sản phẩm dự án tự học của mình, mọi thắc mắc có thể liên hệ mình qua.

    Email: phuctapcode@gmail.com

    Github: [phuccoderr](https://github.com/phuccoderr)
  version: 1.0.0
servers:
  - url: http://localhost:9120/api/v1
    description: User service
  - url: http://localhost:9130/api/v1
    description: Category service
  - url: http://localhost:9140/api/v1
    description: Product service
  - url: http://localhost:9150/api/v1
    description: Customer service
  - url: http://localhost:9160/api/v1
    description: Cart service
  - url: http://localhost:9170/api/v1
    description: Order service
  - url: http://localhost:9180/api/v1
    description: Review service
...
paths:
...
  /users:
    post:
      tags:
        - user-service
      summary: Tạo người dùng quản trị ( Manage )
      description: Tạo người dùng quản trị
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: Thành công
        "400":
          description: Lỗi đầu vào
        "401":
          description: Yêu cầu đăng nhập
        "403":
          description: Quyền không được truy cập
        "422":
          description: Email đã được sử dụng
    get:
      tags:
        - user-service
      summary: Danh sách người dùng quản trị ( Manage )
      description: Danh sách người dùng quản trị
      parameters:
        - name: sort
          in: query
          description: Sort order (asc or desc)
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 10
        - name: keyword
          in: query
          description: Search with keyword
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Thành công
        "400":
          description: Lỗi đầu vào
        "401":
          description: Yêu cầu đăng nhập
        "403":
          description: Quyền không được truy cập
  /users/{id}:
    patch:
      tags:
        - user-service
      summary: Cập nhật người dùng quản trị ( Manage )
      description: Cập nhật người dùng quản trị
      parameters:
        - name: id
          in: path
          description: ID người dùng
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  example: Phuc
                roles:
                  type: array
                  items:
                    type: string
                    example: "ADMIN"
                    enum:
                      - USER
                      - ADMIN
      responses:
        "200":
          description: Thành công
        "400":
          description: Lỗi đầu vào
        "401":
          description: Yêu cầu đăng nhập
        "403":
          description: Quyền không được truy cập
        "404":
          description: Không tìm thấy người dùng
    delete:
      tags:
        - user-service
      summary: Xóa người dùng quản trị ( Manage )
      description: Xóa người dùng quản trị, chỉ có quyền admin mới xoá được!
      parameters:
        - name: id
          in: path
          description: ID người dùng
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Thành công
        "400":
          description: Lỗi đầu vào
        "401":
          description: Yêu cầu đăng nhập
        "403":
          description: Quyền không được truy cập
        "404":
          description: Không tìm thấy người dùng
    get:
      tags:
        - user-service
      summary: Chi tiết người dùng quản trị ( Manage )
      description: Chi tiết người dùng quản trị
      parameters:
        - name: id
          in: path
          description: ID người dùng
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Thành công
        "400":
          description: Lỗi đầu vào
        "401":
          description: Yêu cầu đăng nhập
        "403":
          description: Quyền không được truy cập
        "404":
          description: Không tìm thấy người dùng
  /users/{id}/status/{status}:
    get:
      tags:
        - user-service
      summary: Thay đổi trạng thái người dùng quản trị ( Manage )
      description: Thay đổi trạng thái người dùng quản trị
      parameters:
        - name: id
          in: path
          description: ID người dùng
          required: true
          schema:
            type: string
        - name: status
          in: path
          description: Trạng thái người dùng
          required: true
          schema:
            type: boolean
            enum:
              - true
              - false
            default: true
      responses:
        "200":
          description: Thành công
        "400":
          description: Lỗi đầu vào
        "401":
          description: Yêu cầu đăng nhập
        "403":
          description: Quyền không được truy cập
...
  /customers/upload:
    post:
      tags:
        - customer-service
      summary: Thay ảnh đại diện khách hàng ( Customer )
      description: Thay ảnh đại diện khách hàng
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Thành công
        "401":
          description: Yêu cầu đăng nhập lại
        "404":
          description: Không tìm thấy token
  /customers/account:
    get:
      tags:
        - customer-service
      summary: Tài khoản của tôi ( Customer )
      description: Tài khoản của tôi
      responses:
        "200":
          description: Thành công
        "404":
          description: Không tìm thấy token
    patch:
      tags:
        - customer-service
      summary: Cập nhật tài khoản của tôi ( Customer )
      description: Cập nhật tài khoản của tôi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Thành công
        "401":
          description: Yêu cầu đăng nhập lại
        "404":
          description: Không tìm thấy token
  /customers:
    get:
      tags:
        - customer-service
      summary: Danh sách khách hàng ( Manage )
      description: Danh sách khách hàng
      parameters:
        - name: sort
          in: query
          description: Sort order (asc or desc)
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 10
        - name: keyword
          in: query
          description: Search with keyword
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Thành công
        "400":
          description: Lỗi đầu vào
        "401":
          description: Yêu cầu đăng nhập
        "403":
          description: Quyền không được truy cập
  /customers/{id}:
    get:
      tags:
        - customer-service
      summary: Chi tiết khách hàng ( Manage )
      description: Chi tiết khách hàng
      parameters:
        - name: id
          in: path
          description: ID khách hàng
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Thành công
        "400":
          description: Lỗi đầu vào
        "401":
          description: Yêu cầu đăng nhập
        "403":
          description: Quyền không được truy cập
...
components:
  schemas:
...
    CustomerUpdate:
      properties:
        first_name:
          type: string
          example: Phuc
        last_name:
          type: string
          example: Hoang
        phone_number:
          type: string
          example: 0123456789
        address:
          type: string
          example: Ha Noi
...
    User:
      properties:
        email:
          type: string
          example: phuc@gmail.com
        password:
          type: string
          example: "123456Phuc!"
        name:
          type: string
          example: Phuc
        status:
          type: boolean
          example: true
        roles:
          type: array
          items:
            type: string
            example: "USER"
            enum:
              - USER
              - ADMIN
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Enter JWT token in the format: Bearer {token}"
security:
  - BearerAuth: []